To be is a to be the value of a variable

+++ 0x10-variadic_functions/3-print_all.c
@@ -1,4 +1,4 @@
-#include "variable_functions.h"
+#include "variadic_functions.h"
 #include <stdarg.h>
 #include <stdio.h>
 
@@ -21,16 +21,16 @@ void print_all(const char * const format, ...)
 		{
 			switch (format[i])
 			{
-				case "c":
+				case 'c':
 					printf("%s%c", sep, va_arg(list, int));
 					break;
-				case "i":
+				case 'i':
 					printf("%s%d", sep ,va_arg(list, int));
 					break;
-				case "f":
-					printf("%s%f", sep, va_arg(list, int));
+				case 'f':
+					printf("%s%f", sep, va_arg(list, double));
 					break;
-				case "s":
+				case 's':
 					str = va_arg(list, char *);
 
 					if (!str)
